Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pdfplumber
Pillow
scikit-learn
joblib
python-magic
pandas
pdf2image

transformers==4.35.0
torch

PyMuPDF
python-docx
odfpy